feat: move Staking component to modules folder & fix feedback
