+++ src/routes/Landing.js
@@ -23,7 +23,7 @@ import HomeLayout from "../components/HomeLayout";
 import KText from "../components/KText";
 import Loading from "../components/Loading";
 import { Project } from "./Projects";
-import StakingBox from "../components/Staking/index.tsx";
+import Stake from "../components/Staking/index.tsx";
 
 export default function Landing(props) {
   const [projects, setProjects] = useState([]);
@@ -140,7 +140,7 @@ export default function Landing(props) {
         Staking KULA Token
       </KText>
 
-      <StakingBox {...props} />
+      <Stake {...props} />
 
       <Box mb={40} />
     </>
